1、修改页面布局 2、修改sql字段 3、其他更新

+++ books_manage_provider/src/main/java/com/guaiwuxue/dao/PermissionDao.java
@@ -15,5 +15,5 @@ public interface PermissionDao {
      * @param id
      * @return
      */
-    public Set<Permission> findByRoleId(Integer id);
+    public Set<Permission> findByRoleId(Long id);
 }
